# remove traces from cache after they are sent