Moves some of this logic into the cache so it can be done efficiently. This should lower the memory footprint when the cache capacity is set larger than necessary, and should fix the perf regression associated with repeatedly touching already-sent traces in sendTracesInCache().

The improvement is visible using the benchmark in #131 

```
benchmark                    old ns/op     new ns/op     delta
BenchmarkTraces/batch-8      23715         16990         -28.36%

benchmark                    old allocs     new allocs     delta
BenchmarkTraces/batch-8      286            164            -42.66%
```